requirements.txt
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
google-api-core==2.11.0
google-api-python-client==2.68.0
google-auth==2.15.0
google-auth-httplib2==0.1.0
google-cloud==0.34.0
google-cloud-core==2.3.2
google-cloud-language==2.6.1
google-cloud-speech==2.16.2
google-cloud-storage==2.7.0
google-cloud-translate==3.8.4
google-cloud-videointelligence==2.8.3
google-cloud-vision==3.1.4
google-crc32c==1.5.0
google-resumable-media==2.4.0
googleapis-common-protos==1.57.0
grpcio==1.51.1
grpcio-status==1.51.1
httplib2==0.21.0
idna==3.4
proto-plus==1.22.1
protobuf==4.21.10
pyasn1==0.4.8
pyasn1-modules==0.2.8
pyparsing==3.0.9
requests==2.28.1
rsa==4.9
six==1.16.0
uritemplate==4.1.1
urllib3==1.26.13
